Make size of the table returned by GetComponents() 64
Because there may only ever be 64 concrete component types in the
system (i.e. components of the sort that can be serialized by the
network or used in system entity filters), this seems a reasonable
value, though may be overshot when there are many components of a list
type on the entity.