Change Manifest:replaced back to manifest:updated
Also change pool.onReplace back to pool.onUpdate. "Update" makes more
sense here because we may want component objects that are tables w/ a
proxy table that dispatches the pool's update listeners. In this case
both editing the component in-place as well as outright replacing it
are possible. The signal covers both use cases, but the name "replace"
only reflects one of them.

It may be useful to have facilities in the library to produce Lua
tables with the aforementioned proxy table; investigate