Add Manifest:context
Intended for truly shared data and behavior. This was motivated by the
fact that the Roblox plugin API in may only be accessed by a Script
running as a plugin and cannot be accessed by modules. With this
method, a facade wrapping the plugin API may be stored as a context
variable in the manifest and freely used in modules implementing the
systems.