Add Registry:tryLoad
This method accepts another registry and attempts to load all entities
and components from it. The destination registry must have defined all
the components of the source registry. If any given component fails to
type check, no components of that type will be loaded and a warning
will be printed (as this probably means migration must take place).