Fix for deferred signals
The previous change broke with Workspace.SignalBehavior set to
Deferred (the entries in pendingRemoval/pendingAddition would be set
right back to true at a later invocation point, leading to an infinite
loop of adds and removes). Now, we stop listening to tags during the
main update and explicitly check the entity attribute in the removal
loop.

This change also adds a better error reporting for an unresolvable
component type.